workspaces/python/analytics/pyproject.toml
[project]
name = "analytics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jupyterlab>=4.5.0",
    "pandas>=2.3.3",
    "polars>=1.35.2",
    "scikit-learn>=1.7.2",
    "xgboost>=3.1.2",
    "mlflow>=3.6.0",
    "matplotlib>=3.10.7",
    "seaborn>=0.13.2",
    "plotly>=6.5.0",
    "psycopg2-binary>=2.9.11",
    "sqlalchemy>=2.0.44",
    "connectorx>=0.4.4",
    "onnx>=1.20.0",
    "skl2onnx>=1.19.1",
    "onnxmltools>=1.14.0",
    "python-dotenv>=1.2.1",
]
